fix(checker): run property-level stability detection in the pipeline
Run the request/response property stability checks inside the stability
pipeline, after filterBelowStabilityThreshold rather than as registered
checks. Running after the filter means a property change on an operation
below the configured threshold (already removed from the diff) is not
reported, consistent with how the filter drops below-threshold operations.
A property's stability change never moves its own operation out of scope,
so an in-scope operation survives the filter and its property changes are
still detected. The property rules carry nil handlers (like the endpoint
rules) since the pipeline emits them; GetAllChecks drops to 121.

Report unparseable x-stability-level on a property instead of silently
swallowing it: checkInvalidStabilityLevels only descends into operation
extensions, so a malformed property-level value had no backstop. The
property check now emits api-invalid-stability-level on a parse error,
matching the endpoint behavior.

Reorganize the stability code so each file maps to one concern: the
endpoint checks + shared helpers live in check_api_stability_level.go,
the property checks in check_property_stability_level.go, and checker.go
keeps only the generic pipeline (clone + webhook merge). Tests mirror
the split; the shared stabilityChanges helper moves to helpers_test.go.

Tests assert on the specific check id (containsId) rather than the item
count, matching the established convention and surviving the now-symmetric
request+response emission. Consolidate the find-by-id test helpers into a
single findChange primitive (containsId derives from it; requireChange /
requireNoChange removed). Inline the removeDraftAndAlphaOperationsDiffs
wrapper into applyStabilityLevelPolicy and drop the dead name.
