+++ checker/check_api_stability_level_test.go
@@ -1,7 +1,6 @@
 package checker_test
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/oasdiff/oasdiff/checker"
@@ -9,266 +8,124 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// -------------------------------------------------------
-// Helper: load stability test data files
-// -------------------------------------------------------
+// BC: decreasing stability level is breaking
+func TestBreaking_StabilityLevelDecreased(t *testing.T) {
 
-func getStabilityFile(file string) string {
-	return fmt.Sprintf("../data/stability/%s", file)
-}
-
-func stabilityConfig(sl checker.StabilityLevel, check checker.BackwardCompatibilityCheck) *checker.Config {
-	config := singleCheckConfig(check)
-	config.StabilityLevel = sl
-	return config
-}
-
-// -------------------------------------------------------
-// RequestPropertyStabilityUpdatedCheck tests
-// -------------------------------------------------------
-
-// When StabilityLevel is Beta (default), draft→stable is NOT detected because base (draft) is below threshold
-func TestRequestPropertyStability_BetaLevel_DraftToStableNotDetected(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
+	s1, err := open(deprecationFile("base-beta-stability.yaml"))
 	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-property-all-stable.yaml"))
+
+	s2, err := open(deprecationFile("base-alpha-stability.yaml"))
 	require.NoError(t, err)
 
 	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
 	require.NoError(t, err)
 
-	config := stabilityConfig(checker.StabilityLevelBeta, checker.RequestPropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-	require.Empty(t, errs, "draft→stable should NOT be detected at beta threshold since base (draft) is below threshold")
-}
-
-// When StabilityLevel is Draft, detect draft→stable increase on request property
-func TestRequestPropertyStability_DraftToStable_Increased(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-property-all-stable.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelDraft, checker.RequestPropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
+	// Use Alpha threshold so both beta and alpha are at/above threshold
+	config := allChecksConfig()
+	config.StabilityLevel = checker.StabilityLevelAlpha
+	errs := checker.CheckBackwardCompatibility(config, d, osm)
 	require.Len(t, errs, 1)
-	require.Equal(t, checker.RequestPropertyStabilityIncreasedId, errs[0].GetId())
+
+	require.IsType(t, checker.ApiChange{}, errs[0])
+	e0 := errs[0].(checker.ApiChange)
+	require.Equal(t, checker.APIStabilityDecreasedId, e0.Id)
+	require.Equal(t, "GET", e0.Operation)
+	require.Equal(t, "/api/test", e0.Path)
+	require.Equal(t, "endpoint stability level decreased from `beta` to `alpha`", e0.GetUncolorizedText(checker.NewDefaultLocalizer()))
 }
 
-// When StabilityLevel is Draft, detect stable→draft decrease on request property
-func TestRequestPropertyStability_StableToDraft_Decreased(t *testing.T) {
-	s1, err := open(getStabilityFile("revision-property-all-stable.yaml"))
+// BC: increasing stability level is not breaking
+func TestBreaking_StabilityLevelIncreased(t *testing.T) {
+
+	s1, err := open(deprecationFile("base-alpha-stability.yaml"))
 	require.NoError(t, err)
-	s2, err := open(getStabilityFile("base-property-stable-draft.yaml"))
+
+	s2, err := open(deprecationFile("base-beta-stability.yaml"))
 	require.NoError(t, err)
 
 	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
 	require.NoError(t, err)
+	errs := checker.CheckBackwardCompatibility(allChecksConfig(), d, osm)
+	require.Empty(t, errs)
+}
 
-	config := stabilityConfig(checker.StabilityLevelDraft, checker.RequestPropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
+// BC: specifying an invalid stability level in revision is breaking
+func TestBreaking_InvalidStabilityLevelInRevision(t *testing.T) {
+	s1, err := open(deprecationFile("base.yaml"))
+	require.NoError(t, err)
+
+	s2, err := open(deprecationFile("base-invalid-stability.yaml"))
+	require.NoError(t, err)
+
+	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
+	require.NoError(t, err)
+	errs := checker.CheckBackwardCompatibility(allChecksConfig(), d, osm)
 	require.Len(t, errs, 1)
-	require.Equal(t, checker.RequestPropertyStabilityDecreasedId, errs[0].GetId())
+	require.Equal(t, checker.APIInvalidStabilityLevelId, errs[0].GetId())
+	require.Equal(t, "failed to parse stability level: `value is not one of draft, alpha, beta or stable: \"invalid\"`", errs[0].GetUncolorizedText(checker.NewDefaultLocalizer()))
+	require.Equal(t, "../data/deprecation/base-invalid-stability.yaml", errs[0].GetSource())
 }
 
-// When StabilityLevel is Alpha, draft→stable is NOT reported because base (draft) is below alpha threshold
-func TestRequestPropertyStability_AlphaLevel_DraftToStableNotDetected(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
+// BC: specifying an invalid stability level in base is breaking
+func TestBreaking_InvalidStabilityLevelInBase(t *testing.T) {
+	s1, err := open(deprecationFile("base-invalid-stability.yaml"))
 	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-property-all-stable.yaml"))
+
+	s2, err := open(deprecationFile("base.yaml"))
 	require.NoError(t, err)
 
 	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
 	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelAlpha, checker.RequestPropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-	require.Empty(t, errs, "draft→stable should NOT be reported when base (draft) is below alpha threshold")
-}
-
-// When no property stability actually changed, no changes emitted
-func TestRequestPropertyStability_NoChange(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("base-property-stable-draft.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelDraft, checker.RequestPropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-	require.Empty(t, errs)
-}
-
-// Nil config returns empty
-func TestRequestPropertyStability_NilConfig(t *testing.T) {
-	errs := checker.RequestPropertyStabilityUpdatedCheck(nil, nil, nil)
-	require.Empty(t, errs)
-}
-
-// Nil diffReport.PathsDiff returns empty
-func TestRequestPropertyStability_NilPathsDiff(t *testing.T) {
-	config := checker.NewConfig(checker.GetAllChecks())
-	config.StabilityLevel = checker.StabilityLevelDraft
-	errs := checker.RequestPropertyStabilityUpdatedCheck(&diff.Diff{}, nil, config)
-	require.Empty(t, errs)
-}
-
-// -------------------------------------------------------
-// ResponsePropertyStabilityUpdatedCheck tests
-// -------------------------------------------------------
-
-// When StabilityLevel is Beta (default), draft→stable is NOT detected because base (draft) is below threshold
-func TestResponsePropertyStability_BetaLevel_DraftToStableNotDetected(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-property-all-stable.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelBeta, checker.ResponsePropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-	require.Empty(t, errs, "draft→stable should NOT be detected at beta threshold since base (draft) is below threshold")
-}
-
-// When StabilityLevel is Draft, detect draft→stable increase on response property
-func TestResponsePropertyStability_DraftToStable_Increased(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-property-all-stable.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelDraft, checker.ResponsePropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
+	errs := checker.CheckBackwardCompatibility(allChecksConfig(), d, osm)
 	require.Len(t, errs, 1)
-	require.Equal(t, checker.ResponsePropertyStabilityIncreasedId, errs[0].GetId())
+	require.Equal(t, checker.APIInvalidStabilityLevelId, errs[0].GetId())
+	require.Equal(t, "failed to parse stability level: `value is not one of draft, alpha, beta or stable: \"invalid\"`", errs[0].GetUncolorizedText(checker.NewDefaultLocalizer()))
+	require.Equal(t, "../data/deprecation/base-invalid-stability.yaml", errs[0].GetSource())
 }
 
-// When StabilityLevel is Draft, detect stable→draft decrease on response property
-func TestResponsePropertyStability_StableToDraft_Decreased(t *testing.T) {
-	s1, err := open(getStabilityFile("revision-property-all-stable.yaml"))
+// BC: specifying a non-text, not-json stability level in base is breaking
+func TestBreaking_InvalidNonJsonStabilityLevel(t *testing.T) {
+	s1, err := open(deprecationFile("base.yaml"))
 	require.NoError(t, err)
-	s2, err := open(getStabilityFile("base-property-stable-draft.yaml"))
+
+	s2, err := open(deprecationFile("base-invalid-stability-2.yaml"))
 	require.NoError(t, err)
 
 	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
 	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelDraft, checker.ResponsePropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
+	errs := checker.CheckBackwardCompatibility(allChecksConfig(), d, osm)
 	require.Len(t, errs, 1)
-	require.Equal(t, checker.ResponsePropertyStabilityDecreasedId, errs[0].GetId())
-}
-
-// When StabilityLevel is Alpha, draft→stable is NOT reported because base (draft) is below alpha threshold
-func TestResponsePropertyStability_AlphaLevel_DraftToStableNotDetected(t *testing.T) {
-	s1, err := open(getStabilityFile("base-property-stable-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-property-all-stable.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := stabilityConfig(checker.StabilityLevelAlpha, checker.ResponsePropertyStabilityUpdatedCheck)
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-	require.Empty(t, errs, "draft→stable should NOT be reported when base (draft) is below alpha threshold")
-}
-
-// Nil config returns empty
-func TestResponsePropertyStability_NilConfig(t *testing.T) {
-	errs := checker.ResponsePropertyStabilityUpdatedCheck(nil, nil, nil)
-	require.Empty(t, errs)
-}
-
-// Nil diffReport.PathsDiff returns empty
-func TestResponsePropertyStability_NilPathsDiff(t *testing.T) {
-	config := checker.NewConfig(checker.GetAllChecks())
-	config.StabilityLevel = checker.StabilityLevelDraft
-	errs := checker.ResponsePropertyStabilityUpdatedCheck(&diff.Diff{}, nil, config)
-	require.Empty(t, errs)
+	require.Equal(t, checker.APIInvalidStabilityLevelId, errs[0].GetId())
+	require.Equal(t, "failed to parse stability level: `x-stability-level isn't a string nor valid json`", errs[0].GetUncolorizedText(checker.NewDefaultLocalizer()))
+	require.Equal(t, "../data/deprecation/base-invalid-stability-2.yaml", errs[0].GetSource())
 }
 
 // -------------------------------------------------------
-// Endpoint-level stability change tests (checker.go)
+// Endpoint-level stability change tests
 // -------------------------------------------------------
 
-// When StabilityLevel is Draft, detect endpoint stability decrease (stable→draft)
+// At Draft, an endpoint stability decrease (stable→draft) is reported.
 func TestEndpointStability_StableToDraft_Decreased(t *testing.T) {
-	s1, err := open(getStabilityFile("base-endpoint-stable.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-endpoint-draft.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := allChecksConfig()
-	config.StabilityLevel = checker.StabilityLevelDraft
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-
-	requireChange(t, errs, checker.APIStabilityDecreasedId)
+	errs := stabilityChanges(t, "base-endpoint-stable.yaml", "revision-endpoint-draft.yaml", checker.StabilityLevelDraft)
+	require.True(t, containsId(errs, checker.APIStabilityDecreasedId))
 }
 
-// When StabilityLevel is Draft, detect endpoint stability increase (draft→stable)
+// At Draft, an endpoint stability increase (draft→stable) is reported.
 func TestEndpointStability_DraftToStable_Increased(t *testing.T) {
-	s1, err := open(getStabilityFile("revision-endpoint-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("base-endpoint-stable.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := allChecksConfig()
-	config.StabilityLevel = checker.StabilityLevelDraft
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-
-	requireChange(t, errs, checker.APIStabilityIncreasedId)
+	errs := stabilityChanges(t, "revision-endpoint-draft.yaml", "base-endpoint-stable.yaml", checker.StabilityLevelDraft)
+	require.True(t, containsId(errs, checker.APIStabilityIncreasedId))
 }
 
-// When StabilityLevel is Beta (default), a stable→draft decrease IS detected:
-// the base (stable) is within the threshold, so leaving it is reported. Gating
-// on the lower (draft) level instead would drop this and regress the existing
-// api-stability-decreased ERR.
+// At Beta (default), a stable→draft decrease IS reported: the base (stable) is
+// within the threshold, so leaving it is reported. Gating on the lower (draft)
+// level instead would drop this and regress the existing api-stability-decreased ERR.
 func TestEndpointStability_BetaLevel_StableToDraftDetected(t *testing.T) {
-	s1, err := open(getStabilityFile("base-endpoint-stable.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("revision-endpoint-draft.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := allChecksConfig()
-	// StabilityLevel defaults to Beta
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-
-	requireChange(t, errs, checker.APIStabilityDecreasedId)
+	errs := stabilityChanges(t, "base-endpoint-stable.yaml", "revision-endpoint-draft.yaml", checker.StabilityLevelBeta)
+	require.True(t, containsId(errs, checker.APIStabilityDecreasedId))
 }
 
-// When StabilityLevel is Stable, draft→stable is NOT reported because base (draft) is below threshold
+// At Stable, draft→stable is not reported because base (draft) is below the threshold.
 func TestEndpointStability_StableLevel_DraftToStableNotDetected(t *testing.T) {
-	s1, err := open(getStabilityFile("revision-endpoint-draft.yaml"))
-	require.NoError(t, err)
-	s2, err := open(getStabilityFile("base-endpoint-stable.yaml"))
-	require.NoError(t, err)
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(diff.NewConfig(), s1, s2)
-	require.NoError(t, err)
-
-	config := allChecksConfig()
-	config.StabilityLevel = checker.StabilityLevelStable
-	errs := checker.CheckBackwardCompatibilityUntilLevel(config, d, osm, checker.INFO)
-
-	requireNoChange(t, errs, checker.APIStabilityIncreasedId)
+	errs := stabilityChanges(t, "revision-endpoint-draft.yaml", "base-endpoint-stable.yaml", checker.StabilityLevelStable)
+	require.False(t, containsId(errs, checker.APIStabilityIncreasedId))
 }
