+++ checker/checker.go
@@ -2,17 +2,9 @@ package checker
 
 //go:generate go-localize -input localizations_src -output localizations
 import (
-	"encoding/json"
-	"fmt"
 	"slices"
 
-	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/oasdiff/oasdiff/diff"
-)
-
-const (
-	APIStabilityDecreasedId = "api-stability-decreased"
-	APIStabilityIncreasedId = "api-stability-increased"
 )
 
 // CheckBackwardCompatibility runs the checks with level WARN and ERR
@@ -21,10 +13,7 @@ func CheckBackwardCompatibility(config *Config, diffReport *diff.Diff, operation
 }
 
 // CheckBackwardCompatibilityUntilLevel runs the checks with level equal or higher than the given level.
-// The caller's diffReport is not mutated — internally we clone the
-// PathsDiff and the nested maps/slices that the check pipeline writes
-// to (the rest of the diff is shared by reference, since checks only
-// read it).
+// It does not mutate the caller's diffReport; clonePathsDiffForCheck isolates the pipeline's mutation surface.
 func CheckBackwardCompatibilityUntilLevel(config *Config, diffReport *diff.Diff, operationsSources *diff.OperationsSourcesMap, level Level) Changes {
 	result := make(Changes, 0)
 
@@ -34,7 +23,7 @@ func CheckBackwardCompatibilityUntilLevel(config *Config, diffReport *diff.Diff,
 
 	diffReport = clonePathsDiffForCheck(diffReport)
 
-	result = removeDraftAndAlphaOperationsDiffs(config, diffReport, result, operationsSources)
+	result = applyStabilityLevelPolicy(config, diffReport, result, operationsSources)
 
 	mergeWebhookOperationsIntoPathsDiff(diffReport)
 
@@ -55,209 +44,6 @@ func CheckBackwardCompatibilityUntilLevel(config *Config, diffReport *diff.Diff,
 
 	slices.SortFunc(filteredResult, CompareChanges)
 	return filteredResult
-}
-
-func removeDraftAndAlphaOperationsDiffs(config *Config, diffReport *diff.Diff, result Changes, operationsSources *diff.OperationsSourcesMap) Changes {
-	return applyStabilityLevelPolicy(config, diffReport, result, operationsSources)
-}
-
-// applyStabilityLevelPolicy is a thin pipeline that validates, detects changes, and filters.
-func applyStabilityLevelPolicy(config *Config, diffReport *diff.Diff, result Changes, operationsSources *diff.OperationsSourcesMap) Changes {
-	if diffReport.PathsDiff == nil {
-		return result
-	}
-	result = append(result, checkInvalidStabilityLevels(config, diffReport, operationsSources)...)
-	result = append(result, checkStabilityLevelChanged(config, diffReport, operationsSources)...)
-	filterBelowStabilityThreshold(config, diffReport)
-	return result
-}
-
-// normalizedStability returns STABILITY_STABLE for an empty label, otherwise the label itself.
-func normalizedStability(label string) string {
-	if label == "" {
-		return STABILITY_STABLE
-	}
-	return label
-}
-
-// checkInvalidStabilityLevels emits change records for unparseable x-stability-level values.
-// It does not mutate diffReport.
-func checkInvalidStabilityLevels(config *Config, diffReport *diff.Diff, operationsSources *diff.OperationsSourcesMap) Changes {
-	var result Changes
-
-	for path, pathDiff := range diffReport.PathsDiff.Modified {
-		if pathDiff.OperationsDiff == nil {
-			continue
-		}
-		for operation, operationItem := range pathDiff.OperationsDiff.Modified {
-			if _, err := getStabilityLevel(pathDiff.Base.GetOperation(operation).Extensions); err != nil {
-				result = append(result, getAPIInvalidStabilityLevel(config, operationItem.Base, operationsSources, operation, path, err))
-			}
-			if _, err := getStabilityLevel(pathDiff.Revision.GetOperation(operation).Extensions); err != nil {
-				result = append(result, getAPIInvalidStabilityLevel(config, operationItem.Revision, operationsSources, operation, path, err))
-			}
-		}
-	}
-
-	// Also check deleted paths/operations for invalid stability
-	for _, path := range diffReport.PathsDiff.Deleted {
-		pathVal := diffReport.PathsDiff.Base.Value(path)
-		for operation, operationItem := range pathVal.Operations() {
-			if _, err := getStabilityLevel(pathVal.GetOperation(operation).Extensions); err != nil {
-				result = append(result, getAPIInvalidStabilityLevel(config, operationItem, operationsSources, operation, path, err))
-			}
-		}
-	}
-
-	for path, pathDiff := range diffReport.PathsDiff.Modified {
-		if pathDiff.OperationsDiff == nil {
-			continue
-		}
-		for _, operation := range pathDiff.OperationsDiff.Deleted {
-			operationItem := pathDiff.Base.GetOperation(operation)
-			if _, err := getStabilityLevel(operationItem.Extensions); err != nil {
-				result = append(result, getAPIInvalidStabilityLevel(config, operationItem, operationsSources, operation, path, err))
-			}
-		}
-	}
-
-	return result
-}
-
-// checkStabilityLevelChanged emits APIStabilityDecreasedId / APIStabilityIncreasedId changes.
-// It does not mutate diffReport.
-func checkStabilityLevelChanged(config *Config, diffReport *diff.Diff, operationsSources *diff.OperationsSourcesMap) Changes {
-	var result Changes
-
-	for path, pathDiff := range diffReport.PathsDiff.Modified {
-		if pathDiff.OperationsDiff == nil {
-			continue
-		}
-		for operation, operationItem := range pathDiff.OperationsDiff.Modified {
-			baseStability, err := getStabilityLevel(pathDiff.Base.GetOperation(operation).Extensions)
-			if err != nil {
-				continue // already reported by checkInvalidStabilityLevels
-			}
-			revisionStability, err := getStabilityLevel(pathDiff.Revision.GetOperation(operation).Extensions)
-			if err != nil {
-				continue // already reported by checkInvalidStabilityLevels
-			}
-
-			baseLabel := normalizedStability(baseStability)
-			revisionLabel := normalizedStability(revisionStability)
-			baseLevel := ParseStabilityLevel(baseLabel)
-			revisionLevel := ParseStabilityLevel(revisionLabel)
-
-			if baseLevel == revisionLevel {
-				continue
-			}
-
-			// Gate on the base stability (the level the element is leaving), not the
-			// lower of the two. This reports changes to elements that were within the
-			// configured threshold before the change: a stable->draft destabilization
-			// is reported (base stable meets the threshold), while a draft->stable
-			// change of a previously out-of-scope element is not. Using the lower
-			// level instead would drop stable->draft by default, which regresses the
-			// existing api-stability-decreased ERR.
-			if config != nil && !config.StabilityLevel.IsIncluded(baseLabel) {
-				continue
-			}
-
-			changeId := APIStabilityIncreasedId
-			if revisionLevel < baseLevel {
-				changeId = APIStabilityDecreasedId
-			}
-
-			result = append(result, NewApiChange(
-				changeId,
-				config,
-				[]any{baseLabel, revisionLabel},
-				"",
-				operationsSources,
-				operationItem.Revision,
-				operation,
-				path,
-			))
-		}
-	}
-
-	return result
-}
-
-// filterBelowStabilityThreshold mutates diffReport by removing paths and operations
-// that fall below the configured stability threshold. It emits no Changes.
-func filterBelowStabilityThreshold(config *Config, diffReport *diff.Diff) {
-	if config == nil {
-		return
-	}
-
-	// Filter deleted paths
-	iPath := 0
-	for _, path := range diffReport.PathsDiff.Deleted {
-		keep := false
-		pathVal := diffReport.PathsDiff.Base.Value(path)
-		for operation := range pathVal.Operations() {
-			baseStability, err := getStabilityLevel(pathVal.GetOperation(operation).Extensions)
-			if err != nil {
-				continue
-			}
-			if config.StabilityLevel.IsIncluded(baseStability) {
-				keep = true
-				break
-			}
-		}
-		if keep {
-			diffReport.PathsDiff.Deleted[iPath] = path
-			iPath++
-		}
-	}
-	diffReport.PathsDiff.Deleted = diffReport.PathsDiff.Deleted[:iPath]
-
-	// Filter modified paths → deleted operations and modified operations
-	for _, pathDiff := range diffReport.PathsDiff.Modified {
-		if pathDiff.OperationsDiff == nil {
-			continue
-		}
-
-		// Filter deleted operations
-		iOperation := 0
-		for _, operation := range pathDiff.OperationsDiff.Deleted {
-			operationItem := pathDiff.Base.GetOperation(operation)
-			baseStability, err := getStabilityLevel(operationItem.Extensions)
-			if err != nil {
-				continue
-			}
-			if config.StabilityLevel.IsIncluded(baseStability) {
-				pathDiff.OperationsDiff.Deleted[iOperation] = operation
-				iOperation++
-			}
-		}
-		pathDiff.OperationsDiff.Deleted = pathDiff.OperationsDiff.Deleted[:iOperation]
-
-		// Filter modified operations below threshold
-		for operation := range pathDiff.OperationsDiff.Modified {
-			revisionStability, err := getStabilityLevel(pathDiff.Revision.GetOperation(operation).Extensions)
-			if err != nil {
-				continue
-			}
-			if !config.StabilityLevel.IsIncluded(revisionStability) {
-				delete(pathDiff.OperationsDiff.Modified, operation)
-			}
-		}
-	}
-}
-
-func getAPIInvalidStabilityLevel(config *Config, operation *openapi3.Operation, operationsSources *diff.OperationsSourcesMap, method string, path string, err error) Change {
-	return NewApiChange(
-		APIInvalidStabilityLevelId,
-		config,
-		[]any{err},
-		"",
-		operationsSources,
-		operation,
-		method,
-		path,
-	)
 }
 
 // clonePathsDiffForCheck returns a copy of diffReport whose PathsDiff —
@@ -326,32 +112,4 @@ func mergeWebhookOperationsIntoPathsDiff(diffReport *diff.Diff) {
 	for name, pathDiff := range diffReport.WebhooksDiff.Modified {
 		diffReport.PathsDiff.Modified["webhook:"+name] = pathDiff
 	}
-}
-
-func getStabilityLevel(i map[string]any) (string, error) {
-	if i == nil || i[diff.XStabilityLevelExtension] == nil {
-		return "", nil
-	}
-	var stabilityLevel string
-
-	stabilityLevel, ok := i[diff.XStabilityLevelExtension].(string)
-	if !ok {
-		jsonStabilityRaw, ok := i[diff.XStabilityLevelExtension].(json.RawMessage)
-		if !ok {
-			return "", fmt.Errorf("x-stability-level isn't a string nor valid json")
-		}
-		err := json.Unmarshal(jsonStabilityRaw, &stabilityLevel)
-		if err != nil {
-			return "", fmt.Errorf("failed to unmarshal x-stability-level json")
-		}
-	}
-
-	if stabilityLevel != STABILITY_DRAFT &&
-		stabilityLevel != STABILITY_ALPHA &&
-		stabilityLevel != STABILITY_BETA &&
-		stabilityLevel != STABILITY_STABLE {
-		return "", fmt.Errorf("value is not one of %s, %s, %s or %s: %q", STABILITY_DRAFT, STABILITY_ALPHA, STABILITY_BETA, STABILITY_STABLE, stabilityLevel)
-	}
-
-	return stabilityLevel, nil
 }
