+++ data/stability/base-draft-endpoint-property.yaml
@@ -1,0 +1,21 @@
+openapi: "3.0.0"
+info:
+  title: Test
+  version: "1.0"
+paths:
+  /test:
+    post:
+      operationId: createTest
+      x-stability-level: draft
+      requestBody:
+        content:
+          application/json:
+            schema:
+              type: object
+              properties:
+                f:
+                  type: string
+                  x-stability-level: stable
+      responses:
+        '200':
+          description: OK
